uploader/uploader/src/requirements.txt
bottle
gunicorn
ipython
beaker
webdavclient
